Rename `BaseIndexer` to `ValkyrieIndexer`
